requirements.txt

flask==3.0.3
requests>=2.32.3
google-generativeai>=0.8.3
python-dotenv>=1.0.1
